BZ-1166011 - SessionInfo primary key is Integer
(cherry picked from commit fc9f3f1a977972d1fec600dbe12c2a9b2c49e001)

+++ drools-core/src/main/java/org/drools/core/common/InternalKnowledgeRuntime.java
@@ -22,7 +22,7 @@ public interface InternalKnowledgeRuntime extends KnowledgeRuntime {
 
     InternalProcessRuntime getProcessRuntime();
 
-    void setId(int id);
+    void setId(Long id);
 
     void setEndOperationListener(EndOperationListener listener);
